requirements.txt
websocket-client>=0.54.0